Fixed all build errors I ran into so far. Tested the activity scheduling as well. - NetworkAgent shouldn't return throughput maps of a duration that is different from the ask.   This was happening for the Realtime flows with both uplink and downlink and only capacity is available for one of them. - Session budget calculated for the naive user shouldn't become 0 due to int to double casting. - Added the current auction number update for the non-MC user - If handleRPIresponse doesn't place a bid, handleBidRejection needs to be called to schedule next activity. - Made bidAmount for non-MC user also an integral value. - Non-MC user shouldn't place 0 bids. This throws a run-time error.

+++ src/corenetwork/nodes/Ue.cc
@@ -101,7 +101,7 @@ void Ue::initialize() {
     scriptfilestream.close();
     numOfAuctions = numAc;
 
-    double budgetPerSession = totalBudget/numOfAuctions; // Naive Budget distribution strategy by Non MC User;
+    double budgetPerSession = double(totalBudget)/numOfAuctions; // Naive Budget distribution strategy by Non MC User;
     if (numOfAuctions <= 0) {
         throw cRuntimeError("This user wants nothing ever?");
     }
